pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rainfall-prediction-classifier"
version = "0.1.0"
description = "Portfolio-ready rainfall prediction workflow for Australian weather data."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Repository Maintainer" }
]
dependencies = [
  "joblib>=1.3,<2.0",
  "matplotlib>=3.8,<4.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.1,<3.0",
  "PyYAML>=6.0,<7.0",
  "scikit-learn>=1.4,<2.0",
  "seaborn>=0.13,<1.0",
  "xgboost>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0,<25.0",
  "isort>=5.13,<6.0",
  "pre-commit>=3.6,<4.0",
  "pytest>=8.0,<9.0",
  "ruff>=0.5,<1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100
src_paths = ["src", "scripts", "tests"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["notebooks/*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
